requirements.txt
numpy>=1.21.0
paho-mqtt==1.6.1
pytest>=7.0.0
scipy>=1.7.0
